Fix nasty bug where we set the same hash object to multiple codepoints
This resulted in when we changed one of the codepoint's values, it affected others in the range.
Now we assign the hash to a new variable, and set this new duplicated hash to the codepoint.

This makes UnicodeData() twice as slow, but due to all the other improvements, it only takes about 20 seconds, instead of 10 seconds it made before this fix.