Fix bitfield row offset C function
had an off by one, and was looking up the bitfield row index at the wrong point in the process, so moved that into the get_bitfield_offset function, and only calls that when the letter isn't part of a range.

Now it properly looks up values. Still need to test on high codepoints