Small change but big fix of a bug with missing codepoints
If codepoints had no bitfield row in the %point-index hash, they
are supposed to be assigned to BITFIELD_DEFAULT which is row 0.

There was a bug which caused the codepoint range after an empty
range to sometimes be added to the same range as the empty range.

This also adds in a test for this problem.